requirements.txt
fastapi[standard]
python-dotenv
google-api-core
google-generativeai
IPython
pydantic
Pillow
apscheduler
youtube-transcript-api
pytubefix
pytube
Jinja2
markdown-it-py